RES-3471: move property_tests regression corpus
